Refactor conversation architecture with persistent service connections
